# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "object-storage",
    "description": "Aicacia Object Storage API provides blob services for applications.",
    "contact": { "name": "nathanfaucett@gmail.com" },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "https://spdx.org/licenses/MIT.html"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": ["util"],
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Health check response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Health" }
              }
            }
          },
          "500": {
            "description": "Health check response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Health" }
              }
            }
          }
        }
      }
    },
    "/objects": {
      "get": {
        "tags": ["object"],
        "operationId": "get_objects",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": { "type": ["integer", "null"], "minimum": 0 }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": ["integer", "null"], "minimum": 0 }
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": { "type": ["string", "null"] }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          }
        },
        "security": [{ "Authorization": [] }]
      },
      "post": {
        "tags": ["object"],
        "operationId": "create_object",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateObjectRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          }
        },
        "security": [{ "Authorization": [] }]
      }
    },
    "/objects/by-path": {
      "get": {
        "tags": ["object"],
        "operationId": "get_object_by_path",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          }
        },
        "security": [{ "Authorization": [] }]
      }
    },
    "/objects/by-path/read": {
      "get": {
        "tags": ["object"],
        "operationId": "read_object_by_path",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "", "content": { "*/*": {} } },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          }
        },
        "security": [{ "Authorization": [] }]
      }
    },
    "/objects/{object_id}": {
      "get": {
        "tags": ["object"],
        "operationId": "get_object_by_id",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          }
        },
        "security": [{ "Authorization": [] }]
      },
      "delete": {
        "tags": ["object"],
        "operationId": "delete_object",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "204": { "description": "" },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          }
        },
        "security": [{ "Authorization": [] }]
      }
    },
    "/objects/{object_id}/append": {
      "put": {
        "tags": ["object"],
        "operationId": "append_object",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/UploadPartRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "minimum": 0 }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          }
        },
        "security": [{ "Authorization": [] }]
      }
    },
    "/objects/{object_id}/move": {
      "put": {
        "tags": ["object"],
        "operationId": "move_object",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MoveObjectRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          }
        },
        "security": [{ "Authorization": [] }]
      }
    },
    "/objects/{object_id}/read": {
      "get": {
        "tags": ["object"],
        "operationId": "read_object_by_id",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": { "description": "", "content": { "*/*": {} } },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalError" }
              }
            }
          }
        },
        "security": [{ "Authorization": [] }]
      }
    },
    "/openapi.json": {
      "get": {
        "tags": ["openapi"],
        "operationId": "get_openapi",
        "responses": { "200": { "description": "OpenApi documenation" } }
      }
    },
    "/p2p": {
      "get": {
        "tags": ["p2p"],
        "operationId": "p2p",
        "responses": {
          "200": {
            "description": "P2P response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/P2P" }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HashMap" }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HashMap" }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HashMap" }
              }
            }
          }
        },
        "security": [{ "Authorization": [] }]
      }
    },
    "/version": {
      "get": {
        "tags": ["util"],
        "operationId": "version",
        "responses": {
          "200": {
            "description": "Version response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Version" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateObjectRequest": {
        "type": "object",
        "required": ["path"],
        "properties": {
          "path": { "type": "string" },
          "type": { "type": ["string", "null"] }
        }
      },
      "ErrorMessage": {
        "type": "object",
        "required": ["code", "parameters"],
        "properties": {
          "code": { "type": "string" },
          "parameters": {
            "type": "object",
            "additionalProperties": {},
            "propertyNames": { "type": "string" }
          }
        }
      },
      "ErrorMessages": {
        "type": "array",
        "items": { "$ref": "#/components/schemas/ErrorMessage" }
      },
      "HashMap": {
        "type": "object",
        "additionalProperties": {
          "type": "array",
          "items": { "$ref": "#/components/schemas/ErrorMessage" }
        },
        "propertyNames": { "type": "string" }
      },
      "Health": {
        "type": "object",
        "required": ["db"],
        "properties": { "db": { "type": "boolean" } }
      },
      "InternalError": {
        "type": "object",
        "required": ["status_code", "messages"],
        "properties": {
          "messages": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ErrorMessages"
            },
            "propertyNames": { "type": "string" }
          },
          "status_code": { "type": "integer", "format": "int32", "minimum": 0 }
        }
      },
      "MoveObjectRequest": {
        "type": "object",
        "required": ["path"],
        "properties": {
          "path": { "type": "string" },
          "type": { "type": ["string", "null"] }
        }
      },
      "P2P": {
        "type": "object",
        "required": [
          "enabled",
          "tenant_id",
          "ws_uri",
          "api_uri",
          "id",
          "password"
        ],
        "properties": {
          "api_uri": { "type": "string" },
          "enabled": { "type": "boolean" },
          "id": { "type": "string" },
          "password": { "type": "string" },
          "tenant_id": { "type": "integer", "format": "int64" },
          "ws_uri": { "type": "string" }
        }
      },
      "UploadPartRequest": {
        "type": "object",
        "required": ["part"],
        "properties": {
          "part": {
            "type": "string",
            "format": "binary",
            "contentMediaType": "application/octet-stream"
          }
        }
      },
      "Version": {
        "type": "object",
        "required": ["version", "build"],
        "properties": {
          "build": { "type": "string", "format": "date-time" },
          "version": { "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "Authorization": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "tags": [
    { "name": "object", "description": "Object endpoints" },
    { "name": "util", "description": "Utility endpoints" },
    { "name": "p2p", "description": "P2P endpoints" },
    { "name": "openapi", "description": "OpenApi endpoints" }
  ]
}